requirements.txt
tensorflow
matplotlib
Pillow
kaggle
streamlit
numpy
scipy
opencv-python-headless